Fix: Respect user's first day of week setting in non-Gregorian locales
When the system calendar is not Gregorian (e.g., Chinese, Japanese),
Calendar.solar created a new Gregorian calendar but didn't copy the
user's firstWeekday preference from autoupdatingCurrent.

This caused the calendar to always show Sunday as the first day,
ignoring the user's system preference to start the week on Monday.

Fix: Copy firstWeekday from autoupdatingCurrent when creating
the new Gregorian calendar.

Resolves: Calendar showing Sunday instead of Monday when Monday
is set as first day in System Preferences.